+++ huskarl-core/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Timestamps in JWTs are represented by SystemTime instead of u64.
  - Burn JTI after all other JWT checks have passed.
  - Adds in-memory JTI uniqueness checker (especially for tests).
+ - Add concrete error type (to replace trait).
 
 ## [0.6.4] - 2026-05-29